Use signed bytes.

+++ runtime/src/runtime/opcodes.rs
@@ -287,7 +287,7 @@ impl fmt::Display for Instruction {
 //     substr(tolower($1), 4) " => " i ","
 //   }'
 #[cfg(test)]
-pub fn encode(i: &Instruction) -> u8 {
+pub fn encode(i: &Instruction) -> i32 {
   match i {
     Instruction::Constbyte => 1,
     Instruction::Constshort => 2,
@@ -428,7 +428,7 @@ pub fn encode(i: &Instruction) -> u8 {
 //     print i " => Instruction::"
 //     substr(toupper($1), 3,1) substr(tolower($1), 4) ","
 //   }'
-pub fn decode(i: u8) -> Instruction {
+pub fn decode(i: i32) -> Instruction {
   match i {
     1 => Instruction::Constbyte,
     2 => Instruction::Constshort,
